fix github comments, move wekwebview extension to seperate file, update function name
